혹시 몰라서 학번은 다 Long으로 통일했습니다 (Mysql에서 BIGINT로 자동 매핑)

+++ src/main/java/com/konkukrent/demo/dto/RentalDto/RentalResponseDto.java
@@ -16,7 +16,7 @@ public class RentalResponseDto {
     private String productName;
     private Long userId;
     private String userName;
-    private int studentNum;
+    private Long studentNum;
     private LocalDateTime rentalTime;
     private LocalDate expirationDate;
 }
