+++ src/main/java/com/konkukrent/demo/entity/User.java
@@ -18,7 +18,7 @@ public class User {
     private String userName;
 
     @Column(nullable = false)
-    private int studentNum;
+    private Long studentNum;
 
     @Column(nullable = false, length = 255)
     private String password;
@@ -28,7 +28,7 @@ public class User {
     private Role role;
 
     @Builder
-    private User(String userName, int studentNum, String password, Role role) {
+    private User(String userName, Long studentNum, String password, Role role) {
         this.userName = userName;
         this.studentNum = studentNum;
         this.password = password;
